test(instances): organize failure-path tests by domain
