--- tests/instances/errors/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "errors"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-homepage.workspace = true
-repository.workspace = true
-license.workspace = true
-keywords.workspace = true
-categories.workspace = true
-
-[dependencies]
-rig = { path = "../../rig", features = ["for_tests"] }
-hex = { workspace = true }
-
-[features]
-e2e_proving = ["rig/e2e_proving"]
-cycle_marker = ["rig/cycle_marker"]
